copying: add license and copyright information
The REAME is updated to describe the copyright and license information.
The COPYING file is for GNU GPL v2.
The copyright information for usemem_migrate.c is taken from the header
of migrate_pages.c of the numactl's test directory.
The copyright infromation for usemem.c is taken from Andrew's agreement
email replied to Fengguang.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>