hw_vars: remove the default 300s runtime setting
It could be a surprise for people to find their test cases that doesn't
specify runtime in cmdline to have their run stopped prematurely and
only find the runtime is set in hw_vars.

Tests that have a need to be stopped at 300s should set the runtime
environment variable, unconditionally timeout after 300s in hw_vars
doesn't seem appropriate.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>