usemem: rename the global bytes variable to opt_bytes
The variable bytes is defined as a global one, but some function also
use it as their param. It's difficult to tell if it is the global one
or the local one when one sees bytes. To avoid confusion, rename the
global bytes to opt_bytes.

There should be no functionality change.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>