Base random seed on thread ID instead on PID
The random seed was previously based on process ID. This caused the
random multi-threaded benchmarks to mostly practice very strange
behavior, in which all threads access the same memory, in some
benchmarks in a lockstep behavior.

Changing the getpid to gettid to get the expected behavior, which is
also in line with the multi-process benchmarks.

Signed-off-by: Nadav Amit <namit@vmware.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>