case-shm-pread: use unit_size to control working set
For multi-threaded case, when it needs to exit, these threads will
need to free their memories. Due to the global mmap_sem, the free
has to be done one by one in serial mode. On large servers, due to
the high number of CPU, the free can take more than 1 hour.

To avoid this problem, limit the workingset by letting user to
specify the size. The non-mt case does not need this but it doesn't
hurt(the default size keeps the same as before if unit_size isn't
specified).

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Fengguang Wu <fengguang.wu@intel.com>