hw_vars: limit pagetable to half available memory
We used to use half memory to limit page table, this is OK for setup
that doesn't have other unreclaimable memory consumptions, but consider
if rootfs is hosted on ram and there is no swap, the rootfs consumed
memory is also unreclaimable and can be huge. That can cause some test
OOM.

Solve this by using MemAvailable from /proc/meminfo instead of the
previous MemTotal. This also impacts PTE_LIMIT.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>