case-swap-w-seq: fix ignored --random in case-swap-w-rand-mt
case-swap-w-seq tests for the string -rand in the testcase name to
enable usemem's --random option, but the test breaks when -rand doesn't
appear last, so case-swap-w-rand-mt isn't actually random.

We could search for -rand inside the string, but let's just duplicate
testcase code to avoid similar breakage in the future as more test cases
are added.

Fixes: f35dd1388bfb ("case-swap: add multithreaded test case for case-swap-w-seq/rand")
Acked-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Daniel Jordan <daniel.m.jordan@oracle.com>