case-truncate-seq: truncate sequentially
Mel Gorman suggested LKP to cover a case where sparsefiles are removed
sequentially so add a case based on case-truncate, which did the remove
concurrently.

In the meantime, this script will also report faultin time.

Note the example script provided by Mel in the below link has a 'sync'
after reading of the sparsefile and before removing of it. I failed to
understand the intent of the 'sync' so I omitted.

LINK: https://lore.kernel.org/linux-mm/20220217093113.GU3366@techsingularity.net/
Suggested-by: Mel Gorman <mgorman@techsingularity.net>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>